add new bit for server_id

+++ include/rdma_conn.h
@@ -158,7 +158,7 @@ public:
     struct ibv_mr *m_resp_mr_;
     char *m_reg_buf_;
     struct ibv_mr *m_reg_buf_mr_;
-    uint8_t conn_id_;
+    uint16_t conn_id_;
 
     // << one-sided support >>
     one_side_info m_one_side_info_;
